+++ lint-rules/src/test/kotlin/io/github/santimattius/structured/lint/detectors/CollectAsStateWithoutLifecycleDetectorTest.kt
@@ -76,4 +76,42 @@ class CollectAsStateWithoutLifecycleDetectorTest {
             .run()
             .expectClean()
     }
+
+    @Test
+    fun cleanWhenSuppressLintAnnotationPresent() {
+        val code =
+            """
+            package test
+
+            import android.annotation.SuppressLint
+            import androidx.compose.runtime.Composable
+            import androidx.compose.runtime.collectAsState
+            import kotlinx.coroutines.flow.StateFlow
+
+            class Vm(val ui: StateFlow<Int>)
+
+            @SuppressLint("CollectAsStateWithoutLifecycle")
+            @Composable
+            fun Screen(vm: Vm) {
+                val s by vm.ui.collectAsState()
+            }
+            """.trimIndent()
+
+        val androidAnnotation = """
+            package android.annotation
+            annotation class SuppressLint(vararg val value: String)
+        """.trimIndent()
+
+        TestLintTask.lint()
+            .files(
+                *LintTestStubs.composeRuntimeCollectAndFlow().toTypedArray(),
+                TestFiles.kotlin(androidAnnotation).indented(),
+                TestFiles.kotlin(code).indented(),
+            )
+            .issues(CollectAsStateWithoutLifecycleDetector.ISSUE)
+            .allowMissingSdk()
+            .skipTestModes(TestMode.PARENTHESIZED)
+            .run()
+            .expectClean()
+    }
 }